requirements.txt
torch>=2.0
sentencepiece>=0.1.99
sacrebleu>=2.3
pyyaml>=6.0
tqdm>=4.65
datasets>=2.14
numpy>=1.24
tensorboard>=2.14
